Agent integration

+++ Game/src/Scripts/MainmenuScript.h
@@ -17,6 +17,6 @@ private:
     Entity quitText;
     unsigned int currOption;
     unsigned int numOfOptions;
-    unsigned int cooldownBetweenKeys;
+    float cooldownBetweenKeys;
     float sceneLoadDelay;
 };
